Fix bug whereby login users list wasn't refreshing after cust delete

+++ delete.php
@@ -8,8 +8,9 @@ if(isset($_POST['cust_id'])){
 
     // Delete Customer    
     $cust_id = clean($_POST['cust_id']);
-    delete_addresses_by_cust($cust_id);
-    delete_cust($cust_id);
+    if(delete_addresses_by_cust($cust_id)){
+        delete_cust($cust_id);
+    }
 }
